code/backend/server_py/pyproject.toml
[project]
name = "base-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "casbin>=1.43.0",
    "casbin-async-sqlalchemy-adapter>=1.13.0",
    "dynaconf>=3.2.11",
    "fakeredis>=2.32.1",
    "fastapi>=0.116.1",
    "nuitka>=2.7.16",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.33",
    "numpy>=2.3.2",
    "psutil>=7.0.0",
    "pwdlib[argon2]>=0.2.1",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "sqlite-vec>=0.1.6",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
    # ocr
    "opencv-contrib-python>=4.11.0.86",
    "onnxruntime>=1.22.1",
    "pillow>=11.3.0",
    "shapely>=2.1.1",
    "pyclipper>=1.3.0.post6",
    "freetype-py>=2.5.1",
    "beautifulsoup4>=4.14.2",
    "lxml>=6.0.2",
    "playwright>=1.55.0",
    "brotlicffi>=1.1.0.0",
    "aiofiles>=25.1.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
